# docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      error:
        type: string
    type: object
...
  models.Manufacture:
    properties:
      manufacture_id:
        type: integer
      manufacture_name:
        type: string
    type: object
  models.ResponsePayload:
    properties:
      current_page:
        type: integer
      item_limit:
        type: integer
      items: {}
      next_page:
        type: string
      prev_page:
        type: string
      total_item_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.Status:
    properties:
      status:
        type: string
    type: object
...
info:
  contact:
    email: support@sea-shell.no
    name: API Support
  description: This is the API of GoGear
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoGear API
  version: "1.0"
paths:
...
  /api/v1/manufacture/{manufacture}/delete:
    delete:
      consumes:
      - application/json
      description: Delete manufacture with corresponding ID value
      parameters:
      - description: Unique ID of manufacture you want to update
        in: path
        name: manufacture
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete manufacture with ID
      tags:
      - Manufacture
  /api/v1/manufacture/{manufacture}/get:
    get:
      consumes:
      - application/json
      description: Get manufacture spessific to ID
      parameters:
      - description: Unique ID of manufacture you want to get
        in: path
        name: manufacture
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/models.Manufacture'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get manufacture by ID
      tags:
      - Manufacture
  /api/v1/manufacture/{manufacture}/update:
    post:
      consumes:
      - application/json
      description: Update manufacture identified by ID
      parameters:
      - description: Unique ID of manufacture you want to update
        in: path
        name: manufacture
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Manufacture'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update manufacture with ID
      tags:
      - Manufacture
  /api/v1/manufacture/insert:
    put:
      consumes:
      - application/json
      description: Insert new manufacture with corresponding values
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Manufacture'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Manufacture'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Insert new manufacture
      tags:
      - Manufacture
  /api/v1/manufacture/list:
    get:
      consumes:
      - application/json
      description: Get a list of manufacturers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: search by manufacturename (this is case insensitive and wildcard)
        in: query
        name: manufacture
        type: string
      - description: search by manufactures full name (this is case insensitive and
          wildcard)
        in: query
        name: manufacturename
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.Manufacture'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: List manufacture
      tags:
      - Manufacture
...
securityDefinitions:
  BearerAuth:
    description: Include a server-issued JWT as `Bearer <token>`. Endpoints may require
      either the client or admin audience.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"